build: Support building runtime as a shared library.
This is helpful for building Python bindings.

According to the GYP docs, it also can save a lot of time linking,
so I guess that's a bonus.

This follows the recommendation in the GYP docs:

  `'type'`: This should almost always be set to `<(library)`, which
  allows the user to define at gyp time whether libraries are to be
  built static or shared.

  https://gyp.gsrc.io/docs/UserDocumentation.md

May want to add back some logic to still build a static library in
some circumstances; perhaps by default we ought to build both.

To make this work, also add the `-fPIC` flag at compile time.  Again,
there may be situations where we don't want that, in which case we
ought to add some logic to choose.